rename functions for digitalocean use
